Move Ruby classes from bin/test/ to lib/test/
Rails tried to make these files executable when running `bin/rails
app:update`. That brought to my attention that, anyway, we should not
have these Ruby class definitions in `bin/`, since they aren't intended
to be executable. So, in this change, we move them to `lib/`.
