+++ bin/run-tests
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+bin/pallets --verbose -r ./lib/test/run_pallets_tests.rb
