+++ config/application.rb
@@ -37,7 +37,7 @@ class DavidRunger::Application < Rails::Application
   # Please, add to the `ignore` list any other `lib` subdirectories that do
   # not contain `.rb` files, or that should not be reloaded or eager loaded.
   # Common ones are `templates`, `generators`, or `middleware`, for example.
-  config.autoload_lib(ignore: %w[tasks])
+  config.autoload_lib(ignore: %w[tasks test])
 
   # ActiveJob/Sidekiq
   config.active_job.queue_adapter = :sidekiq
