+++ lib/test/.rubocop.yml
@@ -1,0 +1,4 @@
+inherit_from: ../../.rubocop.yml
+
+Rails/Exit:
+  Enabled: false
