+++ lib/test/diff_helpers.rb
