+++ lib/test/middleware/exit_on_failure_middleware.rb
