+++ lib/test/middleware/task_result_tracking_middleware.rb
