+++ lib/test/run_pallets_tests.rb
@@ -1,10 +1,9 @@
-#!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require_relative 'test/runner.rb'
-require_relative 'test/middleware/exit_on_failure_middleware.rb'
-require_relative 'test/middleware/task_result_tracking_middleware.rb'
-require_relative '../config/initializers/monkeypatches/faraday.rb'
+require_relative '../../config/initializers/monkeypatches/faraday.rb'
+require_relative 'middleware/exit_on_failure_middleware.rb'
+require_relative 'middleware/task_result_tracking_middleware.rb'
+require_relative 'runner.rb'
 
 Pallets.configure do |c|
   c.concurrency = Integer(ENV.fetch('PALLETS_CONCURRENCY', 2))
