+++ lib/test/tasks/build_fixtures.rb
