+++ lib/test/tasks/check_typescript.rb
