+++ lib/test/tasks/compile_java_script.rb
