+++ lib/test/tasks/exit.rb
