+++ lib/test/tasks/pnpm_install.rb
