+++ lib/test/tasks/run_file_size_checks.rb
