+++ lib/test/tasks/run_stylelint.rb
